[IMP] earthoceanfarms: ChartAcccounts Translations
Translating the Mexican Chart of Accounts to English, the original terms
were added as the Spanish Translations. To do it, it was added the
possibility of translating (`translate=True`) the `name` of the
`account.account` model.

A `pre_init_hook` was added to allow the translation of the chart of
accounts in the newly created DB and a migration script to allow the
update in the already created DB.