[FIX] earthoceanfarms: fix currency used to convert capex amount
Previously, the currency used to convert the capex amount was
the company's currency instead of the capex currency.
This caused RFQs with a total amount exceeding 5000 MXN to require
L2 approval, instead of the intended 5000 USD.

With this change, the capex amount will now be converted from
the configured capex currency (USD) to the RFQ currency (MXN).
Consequently, L2 approvers will only need to approve RFQs with a total
amount exceeding 5000 USD.

Related to T#75401