[FIX] earthoceanfarms: fix inherited account.account XML IDs
The account_account_data file modifies account names of an already
installed CoA. Since the CoA is not installed anymore by default,
the XML IDs used in this file are not generated, causing an error
because they are not found.

With this commit, instead of modifying the account.account records,
we modify the account template's names.
By doing so, account names will be the new ones in all installed CoAs,
not only for the main company.

Related to T#75414